added small description to readme

+++ src/main/scala/game/Field.scala
@@ -3,6 +3,6 @@ package game
 /**
  * Created by aleapv on 16.08.2015.
  */
-case class Field(number: Short) {
+case class Field(number: Int) {
 
 }
